+++ integreat_cms/static/src/index.ts
@@ -37,7 +37,8 @@ import "./js/events/event-query-pois.ts";
 import "./js/events/conditional-fields.ts";
 import "./js/events/auto-complete.ts";
 
-import "./js/pages/collapse-subpages.ts";
+import "./js/pages/fetch-subpages.ts";
+import "./js/pages/toggle-subpages.ts";
 import "./js/pages/page-api-token.ts";
 import "./js/pages/page-mirroring.ts";
 import "./js/pages/page-order.ts";