+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 UNRELEASED
 ----------
 
+* [ [#1108](https://github.com/digitalfabrik/integreat-cms/issues/1108) ] Support SVG images in PDF export
+* [ [#1284](https://github.com/digitalfabrik/integreat-cms/issues/1284) ] Inherit status of new translations from source language on XLIFF import
+* [ [#1047](https://github.com/digitalfabrik/integreat-cms/issues/1047) ] Provide option to only export public versions as XLIFF
+* [ [#973](https://github.com/digitalfabrik/integreat-cms/issues/973) ] Support BCP tags for XLIFF import/export
+* [ [#1281](https://github.com/digitalfabrik/integreat-cms/issues/1281) ] Prevent the same push notification from being sent multiple times
+* [ [#760](https://github.com/digitalfabrik/integreat-cms/issues/760) ] Enable linking of push notifications to local news in native apps
+* [ [#1158](https://github.com/digitalfabrik/integreat-cms/issues/1158) ] Prefetch subpages in advance
+* [ [#1052](https://github.com/digitalfabrik/integreat-cms/issues/1052) ] Select all subpages when checking parent page
+* [ [#1004](https://github.com/digitalfabrik/integreat-cms/issues/1004) ] Add button to expand/collapse all pages
+
 
 2022.3.3
 --------